Remove `WriteBlockIndex` in the middle of `ConnectBlock()`
Remove a `WriteBlockIndex` that was in the middle of `ConnectBlock()`. This really shouldn't be in this code at all because the index is written by adding it to a list of dirty. This is very likely what has caused the wallet to be especially likely to corrupt on force close.