Add -debuglogfile option
This patch adds an option to configure the name and/or directory of the debug log.

The user can specify either a relative path, in which case the path is relative to the data directory. They can also specify an absolute path to put the log anywhere else in the file system.

backports https://github.com/bitcoin/bitcoin/commit/cf5f432c69418adb51f0f4a0bdbd6ba3112ee11a