depends: enable lto support for Apple's ld64
Note that this does not _enable_ lto by default in any way, only hooks up the machinery for -flto to work correctly.

enable-lto-support is explicitly used for pinned-clang because we know it works. It is neither enabled nor disabled in the external clang case so that it can be auto-detected.