depends: fix boost mac cross build with clang 9+
The ancient "darwin-4.9.1" profile has long been used to match against clang, which prior to version 9, reported 4.9.1 as its version when invoking "clang++ -dumpversion". Presumably this was a historical compatibility quirk related to Apple's switch from gcc to clang.

This was "fixed" in clang 9.0, so that -dumpversion reports the real version. Unfortunately that had the side-effect of breaking the (brittle) boost compiler detection.

Move to the seemingly more-correct "clang-darwin" profile, which passes the checks and builds correctly.

Also switch to using ar rather than libtool for archiving, as it's what the clang-darwin profile expects to be using.

Note that because this is using a different profile, some of the final command-line arguments end up changing. The changes look sane at a glance.