[Wallet] sort pending wallet transactions before reaccepting
During startup, when adding pending wallet transactions, which spend outputs of other pending wallet transactions, back to the memory pool, and when they are added out of order, it appears as if they are orphans with missing inputs.

Those transactions are then rejected and flagged as conflicting" (= not in the memory pool, not in the block chain).

To prevent this, transactions are explicitly sorted.